requirements.txt
Pillow
pyzbar
pyperclip
pystray
pyperclip